style(racingcar): split by package
