+++ src/main/java/com/nextstep/racingcar/racingcar/Car.java
@@ -1,4 +1,4 @@
-package com.nextstep.racingcar;
+package com.nextstep.racingcar.racingcar;
 
 public class Car {
