+++ src/main/java/com/nextstep/racingcar/racingcar/Racingcar.java
@@ -1,4 +1,4 @@
-package com.nextstep.racingcar;
+package com.nextstep.racingcar.racingcar;
 
 import java.util.*;
