+++ src/main/java/com/nextstep/racingcar/utils/InputCheck.java
@@ -1,4 +1,4 @@
-package com.nextstep.racingcar;
+package com.nextstep.racingcar.utils;
 
 public class InputCheck {
